Added new SQL exercises to README
Two new SQL exercises have been added to the README file. The first exercise involves writing a query that prints the name and price of all phones that have sold more than 5000 units. The second exercise requires writing a query that displays the name and price of specific phone models.